Mild cleanup and details formatting
