+++ docs/asset-manifest.json
@@ -1,13 +1,13 @@
 {
   "files": {
-    "main.css": "/counter/static/css/main.ce963760.css",
-    "main.js": "/counter/static/js/main.5f9994be.js",
+    "main.css": "/counter/static/css/main.adcdb9f8.css",
+    "main.js": "/counter/static/js/main.4fa00f53.js",
     "index.html": "/counter/index.html",
-    "main.ce963760.css.map": "/counter/static/css/main.ce963760.css.map",
-    "main.5f9994be.js.map": "/counter/static/js/main.5f9994be.js.map"
+    "main.adcdb9f8.css.map": "/counter/static/css/main.adcdb9f8.css.map",
+    "main.4fa00f53.js.map": "/counter/static/js/main.4fa00f53.js.map"
   },
   "entrypoints": [
-    "static/css/main.ce963760.css",
-    "static/js/main.5f9994be.js"
+    "static/css/main.adcdb9f8.css",
+    "static/js/main.4fa00f53.js"
   ]
 }
