+++ docs/index.html
@@ -1,1 +1,1 @@
-<!doctype html><html lang="en"><head><meta charset="utf-8"/><link rel="icon" href="/counter/favicon.ico"/><meta name="viewport" content="width=device-width,initial-scale=1"/><meta name="theme-color" content="#000000"/><meta name="description" content="Kyle and Ben's Take Counter"/><link rel="apple-touch-icon" href="/counter/logo192.png"/><link rel="manifest" href="/counter/manifest.json"/><title>Take 1</title><script defer="defer" src="/counter/static/js/main.5f9994be.js"></script><link href="/counter/static/css/main.ce963760.css" rel="stylesheet"></head><body><noscript>You need to enable JavaScript to run this app.</noscript><div id="root"></div></body></html>
+<!doctype html><html lang="en"><head><meta charset="utf-8"/><link rel="icon" href="/counter/favicon.ico"/><meta name="viewport" content="width=device-width,initial-scale=1"/><meta name="theme-color" content="#000000"/><meta name="description" content="Kyle and Ben's Take Counter"/><link rel="apple-touch-icon" href="/counter/logo192.png"/><link rel="manifest" href="/counter/manifest.json"/><title>Take 1</title><script defer="defer" src="/counter/static/js/main.4fa00f53.js"></script><link href="/counter/static/css/main.adcdb9f8.css" rel="stylesheet"></head><body><noscript>You need to enable JavaScript to run this app.</noscript><div id="root"></div></body></html>
