--- src/Details.js
@@ -1,10 +1,0 @@
-import { useState, useEffect, useCallback } from 'react';
-
-export default function Details() {
-
-    return <div class="d-flex justify-content-start border-bottom align-items-baseline">
-                <h3 class="m-2">Client: </h3>
-                <input type='text' class="w-100 h3 text-white bg-transparent border-0" />
-    </div>
-
-}
