+++ src/components/Details.js
@@ -1,0 +1,18 @@
+export default function Details() {
+
+    return <div className="fixed-top">
+             <div className="container mt-2 pl-2 pr-2">
+               <div className="row no-gutters no-wrap justify-content-start align-items-stretch">
+                 <div className="col-auto p-0 border-bottom">
+                   <h3 className="m-2">Client: </h3>
+                 </div>
+                 <div className="col-auto col-sm-6 p-0 border-bottom">
+                   <input type='text' className="w-100 h3 text-white bg-transparent h-100 border-0" style={{outline: "none"}} />
+                 </div>
+                 <div className="col d-none d-sm-block">
+                 </div>
+               </div>
+             </div>
+           </div>
+
+}
