+++ src/util.js
@@ -1,0 +1,8 @@
+export function setMinimum(fn, min=1) {
+  return (arg) => {
+    if (arg < min) {
+      arg = min;
+    }
+    fn(arg);
+  }
+}
